lesson/requirements.txt
certifi==2019.9.11
Django==2.2.7
django-cors-headers==3.2.1
django-filter==2.3.0
djangorestframework==3.10.3
pytz==2019.3
sqlparse==0.3.0
